# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "In Abstract Controller",
        "version": "1"
    },
    "paths": {
...
        "/users/auth/registeration/phone-number-step": {
            "post": {
                "tags": [
                    "usersAuth"
                ],
                "operationId": "aefe4cf7655ae7cc7f4a98206a67d6c4",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddPhoneNumberRegisterationStepRequestData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item fetched successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddPhoneNumberRegisterationStepResponseData"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AddPhoneNumberRegisterationStepRequestData": {
                "properties": {
                    "phone_number": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "AddPhoneNumberRegisterationStepResponseData": {
                "properties": {
                    "token": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "ref_to_be_used_below_csrf": {
                "type": "apiKey",
                "name": "X-XSRF-TOKEN",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ref_to_be_used_below_csrf": []
        }
    ],
...
}